+++ tests/core-addon/integration/core_addon.js
@@ -86,13 +86,33 @@ describe("Core-Addon Onboarding", function () {
     await findAndAct(this.driver, By.css(`[label="Add"]`), e => e.click());
     await findAndAct(this.driver, By.css(`[label="Okay, Got It"]`), e => e.click());
 
-    // FIXME close tab and click on icon, check that post-enrollment options page is shown.
-    // This will currently fail because there is a bug in the core-addon UI, where
-    // the options page will show no studies.
-    // See https://github.com/mozilla-rally/core-addon/issues/235
+    // Close options page tab.
+    await this.driver.close();
+    const originalTab = (await this.driver.getAllWindowHandles())[0];
+    await this.driver.switchTo().window(originalTab);
+    await findAndAct(this.driver, By.id("rally-core_mozilla_org-browser-action"), e => e.click());
+
+    // We expect the extension to load its options page in a new tab.
+    console.debug("debug1");
+    await this.driver.wait(async () => {
+      console.debug("window handles:", (await this.driver.getAllWindowHandles()).length);
+      return (await this.driver.getAllWindowHandles()).length >= 2;
+    }, WAIT_FOR_PROPERTY);
+    console.debug("debug2");
+
+    // Selenium is still focused on the old tab, so switch to the new window handle.
+    let latestTab = (await this.driver.getAllWindowHandles()).length - 1;
+    const newOptionsTab = (await this.driver.getAllWindowHandles())[latestTab];
+    await this.driver.switchTo().window(newOptionsTab);
 
     // Switch back to web content context.
     await this.driver.setContext(firefox.Context.CONTENT);
+
+    console.debug("debug3");
+    // Ensure that the study card for the base study is displayed.
+    const baseStudySelector = By.xpath(`//span[text()="Ion Base Study"]`);
+    await this.driver.findElement(baseStudySelector);
+    console.debug("debug4");
 
     // Begin study unenrollment cancel it.
     await findAndAct(this.driver, By.xpath(`//button[text()="Leave Mozilla Rally"]`), e => e.click());